aec_app/requirements.txt
toga-core==0.5.2
toga-android==0.5.2
numpy==1.26.2
pyaec==0.1.0b1
